requirements.txt
django==4.1
Pillow==9.2.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
djangorestframework-simplejwt==5.2.0
django-debug-toolbar==3.6.0
